+++ priv/repo/migrations/20230508121000_create_events_event_index.exs
@@ -1,0 +1,7 @@
+defmodule TdAudit.Repo.Migrations.CreateEventsEventIndex do
+  use Ecto.Migration
+
+  def change do
+    create index("events", [:event])
+  end
+end