+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-5711] Create index for Events `event` column and change `self_reported_event_type` to single value instead of list
+
 ## [5.5.0] 2023-04-18
 
 ### Added